Remove tests from excluded files in tsconfig too fix Webstorm type errors on HTMLElements.
The previous commit to change the order of the lib array only worked until quitting webstorm. Found a comment on Jetbrains' YouTrack saying the tsconfig only applies to files in the "include" array, and the default configuration doesn't include the dom types. I checked that the test files are not included in the app bundle when running locally, and will double-check they also aren't included on heroku after this commit is deployed